move NpcHologramLoop listener into HologramListener from HologramProvider
